fix popover outside click
